Move statusbar modules and evaluate to api
This decouples the actual bar object and the intelligence of the status
messages and allows using the messages elsewhere.
